+++ config/config.go
@@ -103,21 +103,6 @@ type Config struct {
 	lastFileModTime time.Time // last modification time
 }
 
-// toPtr returns a pointer to the given value
-func toPtr[T any](value T) *T {
-	return &value
-}
-
-// contains checks if a key exists in the slice
-func contains[T comparable](slice []T, key T) bool {
-	for _, item := range slice {
-		if item == key {
-			return true
-		}
-	}
-	return false
-}
-
 // GetMetricLabels returns a slice of unique metric labels from all enabled endpoints
 // in the configuration. It iterates through each endpoint, checks if it is enabled,
 // and then collects unique labels from the endpoint's labels map.
